+++ MeliMessageSender/Controllers/QueueService.cs
@@ -1,0 +1,19 @@
+using Microsoft.WindowsAzure.Storage.Queue;
+
+namespace MeliMessageSender.Controllers
+{
+	public class QueueService
+	{
+		private readonly CloudQueue _cloudQueue;
+
+		public QueueService(CloudQueue cloudQueue)
+		{
+			_cloudQueue = cloudQueue;
+		}
+
+		public virtual void AddMessage(CloudQueueMessage message)
+		{
+			_cloudQueue.AddMessage(message);
+		}
+	}
+}
